requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alpaca-py==0.42.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
azure-core==1.36.0
azure-identity==1.25.1
azure-storage-blob==12.27.0
bandit==1.8.6
beautifulsoup4==4.14.2
black==25.9.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
curl_cffi==0.13.0
deprecation==2.1.0
fastapi==0.119.1
frozendict==2.4.6
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.0.3
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.9.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pandas-ta==0.4.71b0
pathspec==0.12.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.5
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.1
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
ruff==0.14.1
scipy==1.16.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sseclient-py==1.8.0
starlette==0.48.0
stevedore==5.5.0
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==13.0
xgboost==3.1.1
yarl==1.22.0
yfinance==0.2.66
